requirements.txt
google-adk
litellm
openai
pydantic
